+++ src/services/api.js
@@ -779,3 +779,11 @@ export function unblockUserInGroup({ groupName, username }) {
 export function sendVerificationCode({ email, mode }) {
   return fetch(`${apiRoot}/v2/users/verifyEmail`, postRequestOptions('POST', { email, mode }));
 }
+
+export function disableBansInGroup({ groupName }) {
+  return fetch(`${apiRoot}/v2/groups/${groupName}/disableBans`, postRequestOptions());
+}
+
+export function enableBansInGroup({ groupName }) {
+  return fetch(`${apiRoot}/v2/groups/${groupName}/enableBans`, postRequestOptions());
+}